I **think** that there was **technically** a race condition here.
At the minimum, it looks bad to make naked writes to a struct field that's larger than the size of any real-world machine's native integer size from multiple threads at once, especially when we can just burn seed 0 to completely eliminate the problem.

I could probably add an Interlocked.CompareExchange and a Volatile.Read in there to get this "perfect", but I don't want to.