#35 (and #47 to an extent): Initial support for facing the player in the direction of its target.
This isn't perfect: ideally, when the current and target location are the same, we should angle it from the location that the player came from rather than just choosing not to change the angle at all.

That said, in the cases where we REALLY want to know where the player is going — i.e., when it's going a really long distance — this will be more than sufficient.