Deal with an unmanaged memory leak, as best I can.
It seems that something in Avalonia's image rendering code may have a leak, since we kept giving it the same Bitmap objects back and forth and unmanaged memory usage grew seemingly without bound.

The workaround for now: instead of flipping back and forth between which Bitmap object we use for a single Image, put two Images in the view and flip IsVisible back and forth.

Resolves #98