Change this to "hard locked" and "soft locked" regions.
"Hard locked" = it is impossible to get to that location because all paths that would lead to it involve at least one landmark whose requirements have not yet been met.

"Soft locked" = it is not "hard locked", but all paths that lead to it involve at least one unchecked landmark whose requirements have been met.

For the time being, "soft locked" is pretty much write-only, but the distinction will be needed when "Startled" is no longer as special a special-case.