*PROBABLY* fix what was *PROBABLY* a race condition related to the pause button.
Symptom: on (perhaps rare) occasions, clicking the Pause button would only pause animations; the player token would still move around making checks.

If I'm not mistaken, this *should* keep the game more-or-less consistent the state of the Pause button. The button will still probably never perfectly immediately freeze everything with millisecond-level precision, but you're not looking for that, so I don't need to overcomplicate it.