Some wholly unnecessary, medium-risky, and probably poorly-decided refactoring that may or may not help with #52 and #67 when I get to it.
The primary aim, if this can be said to have any, was to have everything consistently settle on one preferred standard for how to reference items, locations, and regions, instead of having some pointers to the objects themselves, some keys to locations' positions within their parent regions, some even just a u16 number.

Now:
- region keys are u8
- location and item keys are u16
- there's also a separate location key that's a pair of u8 values like before
- everyone is expected to resolve their object by GameDefinitions.Instance[theirKey], which is a trivial pointer lookup into an array.

As I went through this, I focused on making sure that the code takes full advantage of the better data layout that this enables (it's reminiscent of a project I did at my day job), because I kept waking up too early and wanting something mindless to do.