Escape regex characters when attempting to match a chat command.
Normalize Unicode strings so that "@Stéve" won't fail to match just because the slot name is actually "Stéve".

Allow matching by alias as well, just in case someone generates a world with two slots of canonically equivalent names and wants to direct one rat independently of the other.

Resolves #122