Start a file for determining stuff related to reaching target locations.
Also do some routine version updates and switch to `@typescript/native-preview` since WebStorm supports that now. This demands some imperfections in `game-tab-text-client.ts` in order to work around some deficiencies in the current preview, but it's not SO bad. For now.

Also mark the CJS modules that seem acceptable to leave alone today: one because it's trivially tiny anyway, and the other because it LOOKS like it's only used from one place that we don't use anyway.