Let's take a bit of a detour here.
The thing I need to prove out right now is not whether or not the GUI can update, but that I can port the actual game logic to TypeScript in a way that makes sense.

That's what's sticking me, so let's strip away the GUI bits and just hard focus on that.