fix an oversight: we can't JUST tweak the "get target location" code to cover for a substandard implementation of "get route to target location".
there could be multiple different legal paths to the target, so if we're going to stop at the first unchecked landmark on the way, then we can't just pick any route.

fortunately, this isn't going to kill performance: we touch all reachable locations during the target location calculation anyway, and we already need to mark which ones are touched in an array, so we can actually build this "prev" array on-the-spot.

probably, some comments are now invalid that I don't want to look for. oh well.