Replace Pixi.js with browser-native elements (#132)
Pixi.js v8 relies on WebGL or WebGPU, so some users get left behind.

This kind of application just isn't intense enough to merit such an extreme response, and the fact that we need browser-native elements anyway for reasonable tooltips is just more evidence that this is not appropriate.

Moving to browser-native elements also allows for a smoother animation story using the Web Animations API, since each animation has a .finished that gives a Promise which will be resolved (or, if canceled, rejected) when the animation is finished, allowing for a chain of actions to be animated without trickery.